Move wasm impl into PodElGamalPubkey
