--- zk-sdk/src/encryption/wasm/mod.rs
@@ -1,1 +1,0 @@
-pub mod elgamal;
