+++ zk-sdk/src/lib.rs
@@ -25,8 +25,6 @@ pub mod pod;
 mod range_proof;
 mod sigma_proofs;
 mod transcript;
-#[cfg(all(not(target_os = "solana"), target_arch = "wasm32"))]
-pub mod wasm;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod zk_elgamal_proof_program;
